upgrade services to Bookings::Creator
